# Conflicts:
#	tetris_player/src/Main.cpp
#	tetris_player/src/include/FileWatcher.hpp
#	tetris_player/src/include/Logger.hpp
#	tetris_player/src/include/TetrisSolver/PlayState.hpp
#	tetris_player/src/include/TetrisSolver/Tetrimino.hpp
#	tetris_player/src/include/TetrisSolverSerial.hpp

+++ tetris_player/src/include/TetrisSolver/PlayState.hpp
@@ -34,7 +34,7 @@ public:
 	 * @param column Column of the play_area at which to attempt to add the tetrimino
 	 * @return Whether place was successful
 	 */
-	bool place(Tetrimino::Figure figure, int rotation, size_t column);
+	bool place(Tetrimino::Figure figure, size_t rotation, size_t column);
 
 	/**
 	 * @brief Returns a value according to the pieces placed on the game board
@@ -45,7 +45,7 @@ public:
 	unsigned int score() const;
 
 	/**
-	 * @brief Returns a game state identifier 
+	 * @brief Returns a game state identifier
 	 * @details Get the identifier of the current game state
 	 * @return id: int
 	 *
@@ -95,9 +95,9 @@ private:
  * @param column Column of the play_area at which to attempt to add the tetrimino
  * @return Whether place was successful
  */
-bool PlayState::place(Tetrimino::Figure figure, int rotation, size_t column)
+bool PlayState::place(Tetrimino::Figure figure, size_t rotation, size_t column)
 {
-	const Tetrimino& tetrimino{Tetrimino::getTetrimino(rotation, figure)};
+	const Tetrimino& tetrimino{ Tetrimino::getTetrimino(rotation, figure) };
 
 	if (tetrimino.width > playArea.cols - column)
 	{
@@ -166,16 +166,20 @@ bool PlayState::place(Tetrimino::Figure figure, int rotation, size_t column)
  */
 unsigned int PlayState::score() const
 {
-	unsigned int score{0};
+	unsigned int score{ 0 };
 
-	for (size_t i{0}; i < playArea.rows; ++i) {
-		unsigned int sum{0};
-		for (size_t j{0}; j < playArea.cols; ++j) {
-			if (playArea[i][j] != '0') {
+	for (size_t i{ 0 }; i < playArea.rows; ++i)
+	{
+		unsigned int sum{ 0 };
+		for (size_t j{ 0 }; j < playArea.cols; ++j)
+		{
+			if (playArea[i][j] != '0')
+			{
 				++sum;
 			}
 		}
-		if (sum > 0) {
+		if (sum > 0)
+		{
 			score += sum * i * i;
 		}
 	}
